better srp and pfl impl
