+++ Dockerfile
@@ -8,4 +8,3 @@ RUN apk update
 RUN apk add gcc g++ gdb git cmake make python3 linux-headers
 
 COPY . .
-
