--- libmem/include/libmem/libmem.hpp
@@ -1,34 +1,0 @@
-/*
- *  ----------------------------------
- * |         libmem - by rdbo         |
- * |      Memory Hacking Library      |
- *  ----------------------------------
- */
-
-/*
- * Copyright (C) 2022    Rdbo
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-#pragma once
-#ifndef LIBMEM_HPP
-#define LIBMEM_HPP
-
-#ifndef LM_FORCE_LANG_CPP
-#define LM_FORCE_LANG_CPP
-#endif
-
-#include "libmem.h"
-
-#endif /* LIBMEM_HPP */
