+++ src/asm.c
