+++ src/internal.h
