+++ src/internal/win/handle.c
