+++ src/scan.c
