+++ src/symbol.cpp
