+++ src/vmt.c
