+++ tools/docker-env.sh
@@ -1,6 +1,5 @@
 #!/bin/sh
-
+cd ../
 mkdir -p build
 docker build -t libmem .
 docker run -v "$(pwd)/build:/app/build" -it libmem
-
